Switch deal models and calculations to Decimal type
Replaces float fields with Decimal in Pydantic models, calculation logic, and SQLAlchemy models for improved precision in financial computations. Updates all relevant request/response schemas, calculation functions, and database columns to use Decimal/Numeric types instead of float. This change ensures more accurate handling of monetary values and percentages throughout the backend.

+++ BackEnd/ReqRes/analyzeFlip/analyzeFlipRes.py
@@ -1,13 +1,14 @@
 from typing import Optional
+from decimal import Decimal
 from pydantic import BaseModel
 
 class analyzeFlipRes(BaseModel):
     """Represents the result of the Flip calculation."""
-    net_profit: float
-    roi: float
-    annualized_roi: float
-    total_cash_needed: float
-    total_holding_costs: float
-    total_hml_interest: float
+    net_profit: Decimal
+    roi: Decimal
+    annualized_roi: Decimal
+    total_cash_needed: Decimal
+    total_holding_costs: Decimal
+    total_hml_interest: Decimal
     messages: Optional[list[str]] = None
